- Add postgres tables - Rename index on sqlite3
